Fulfills bonus .forEach() requirement
Requirement states that the callback function should not be called
when there is a hole in the array. This partially meets the edge
case requirement called by the MDN .forEach() spec that states
that the callback funciton should not be run on index properties
that have been deleted or are uninitialized i.e. sparse arrays